+++ src/components/audit/AuditBlock.vue
@@ -1,5 +1,5 @@
 <template>
-  <BasicBlock title="Logs" icon="pi-book">
+  <AppContainer title="Logs" icon="pi-book">
     <DataTable :value="store.dashboardEntries" data-key="id" :loading="store.loading" class="my-3">
       <Column field="createdAt" header="Timestamp">
         <template #body="slotProps">
@@ -10,12 +10,12 @@
       <Column field="action" header="Action" />
     </DataTable>
     <RouterLink to="/audit">More recent activity</RouterLink>
-  </BasicBlock>
+  </AppContainer>
 </template>
 
 <script setup lang="ts">
 import { useAuditStore } from '@/stores/audit.store';
-import BasicBlock from '@/components/BasicBlock.vue';
+import AppContainer from '@/layout/AppContainer.vue';
 const store = useAuditStore();
 store.setTake(5);
 store.setSkip(0);
