+++ src/views/Infoscreen/SettingsView.vue
@@ -1,26 +1,24 @@
 <template>
-  <div class="page-container">
-    <div class="page-title">Settings</div>
-    <div class="flex flex-column gap-5">
-      <div class="flex justify-content-between">
-        <div style="width: 30%">
-          <TimeSettings />
-        </div>
-        <div style="width: 30%">
-          <RoomResponsibleSettings />
-        </div>
-        <div style="width: 30%">
-          <EROSettings />
-        </div>
+  <AppContainer title="Settings" icon="pi-cog">
+    <div class="grid">
+      <div class="col-4">
+        <TimeSettings />
+      </div>
+      <div class="col-4">
+        <RoomResponsibleSettings />
+      </div>
+      <div class="col-4">
+        <EROSettings />
       </div>
     </div>
-  </div>
+  </AppContainer>
 </template>
 
 <script setup lang="ts">
 import TimeSettings from '@/components/infoscreen/TimeSettings.vue';
 import RoomResponsibleSettings from '@/components/infoscreen/RoomResponsibleSettings.vue';
 import EROSettings from '@/components/infoscreen/EROSettings.vue';
+import AppContainer from '@/layout/AppContainer.vue';
 </script>
 
 <style scoped></style>
