+++ src/views/Modes/CenturionModeView.vue
@@ -1,6 +1,5 @@
 <template>
-  <div class="page-container">
-    <div class="page-title">Centurion Mode</div>
+  <AppContainer title="Centurion Mode" icon="pi-crown">
     <div class="grid">
       <div class="col-12 md:col-6 lg:col-5">
         <CenturionTapeSelect />
@@ -9,7 +8,7 @@
         <CenturionTapeActive />
       </div>
     </div>
-  </div>
+  </AppContainer>
 </template>
 
 <script setup lang="ts">
@@ -17,6 +16,7 @@ import { useCenturionStore } from '@/stores/modes/centurion.store';
 import { onUnmounted } from 'vue';
 import CenturionTapeSelect from '@/components/modes/centurion/CenturionTapeSelect.vue';
 import CenturionTapeActive from '@/components/modes/centurion/CenturionTapeActive.vue';
+import AppContainer from '@/layout/AppContainer.vue';
 
 const store = useCenturionStore();
 store.init();
